requirements.txt
typer>=0.9.0
opencv-python>=4.8.0
pathlib
numpy>=1.21.0
PyYAML>=6.0